Fit Bayesian Decision Curve Analysis using Stan for list of models or binary tests

Fit Bayesian Decision Curve Analysis using Stan for list of models or binary tests

Usage

.dca_stan_list(
  n_thr,
  n_models_or_tests,
  N,
  d,
  tp,
  tn,
  thresholds,
  prior_p1,
  prior_p2,
  prior_Se1,
  prior_Se2,
  prior_Sp1,
  prior_Sp2,
  N_ext,
  d_ext,
  other_models_indices,
  refresh = 0,
  ...
)

Arguments

n_thr

Number of thresholds (int.).

n_models_or_tests

Number of models or binary tests (int.).

N

Sample size (vector of integers of length n_thr).

d

Diseased: number of diseased persons or events (vector of integers of length n_thr).

tp

True Positives: number of diseased persons correctly identified as such by the diagnostic test of prediction model (matrix of integers of size n_thr by n_models_or_tests).

tn

True Negatives: number of diseased persons correctly identified as such by the diagnostic test of prediction model (matrix of integers of size n_thr by n_models_or_tests).

thresholds

Numeric vector with probability thresholds with which the net benefit should be computed (default is seq(0.01, 0.5, 0.01)).

N_ext, d_ext

External sample size and number of diseased individuals (or cases), respectively, used to adjust prevalence.

refresh

Control verbosity of rstan::sampling.

...

Arguments passed to rstan::sampling (e.g. iter, chains).

prior_p, prior_se, prior_sp

Prior parameters for prevalence, sensitivity, and specificity (numeric matrices of size n_thr by n_models_or_tests).

Value

An object of class stanfit returned by rstan::sampling (e.g. iter, chains)